// app/src/components/common/enums/DeviceDataProperty.ts
export enum DeviceDataProperty {
  BARRIER_ID = "barrierId",
  MESSAGE_TYPE = "messageType",
  BATTERY = "battery",
  PROFILE = "profile",
  CMD_ACK = "cmdAck",
  GPS_FIX_AGE = "gpsFixAge",
  SAT_IN_FIX = "satInFix",
  LAT = "lat",
  LNG = "lng",
  BATTERY_LIFE_COUNT = "batteryLifeCount",
  BATTERY_LIFE = "batteryLife",
  FCNT_UP = "fCntUp",
  BARRIER_MESSAGE_BATCH = "barrierMessageBatch",
  META = "meta",
  DOWNLINK_PAYLOAD = "downlinkPayload",
  BARRIER_OPEN_CLOSE_TIME_BATCH = "barrierOpenCloseTimeBatch",
  REBOOT_REASON = "rebootReason",
  M1_M2_ID = "m1M2ID",
  SRC_ID = "srcID",
  LINE_NR = "lineNR",
  VERSION = "version",
  LO_LO_TEMP = "loLoTemp",
  LO_TEMP = "loTemp",
  HI_TEMP = "hiTemp",
  HI_HI_TEMP = "hiHiTemp",
  LO_LO_EXT_TEMP = "loLoExtTemp",
  LO_EXT_TEMP = "loExtTemp",
  HI_EXT_TEMP = "hiExtTemp",
  HI_HI_EXT_TEMP = "hiHiExtTemp",
  BAROM_BAR = "baromBar",
  INT_TEMP = "intTemp",
  INT_HUM = "intHum",
  EXT_TEMP = "extTemp",
  EXT_HUM = "extHum",
  V_BAT = "vBat",
  ID = "id",
  TIME_RECEIVED = "timeReceived",
  TIME_SET = "timeSet",
  QUERY = "query",
  END_TIME = "endTime",
  PERIOD = "period",
  START_TIME = "startTime",
  VALUE = "value",
}